modify configure rootheight type and networkid type (#81)
* modify configure based on 2.4.0

* add deepCopy func

* add getShardConfig func

* modify the NewShardConfig, use ChainConfig = chainCfg, use reference instead of copy.

* modify grpc test to depend on 2.4.0 version of cfg

* rm unuseful json tag

* modify address from string to account.Address

* modify config module

* modify root height type and networkid type

* modify rootheight type in ShardGenesis struct

* modify rootheight and about this operation

+++ cluster/config/shard_config.go
@@ -8,7 +8,7 @@ import (
 )
 
 type ShardGenesis struct {
-	RootHeight         uint64                       `json:"ROOT_HEIGHT"`
+	RootHeight         uint32                       `json:"ROOT_HEIGHT"`
 	Version            uint32                       `json:"VERSION"`
 	Height             uint64                       `json:"HEIGHT"`
 	HashPrevMinorBlock string                       `json:"HASH_PREV_MINOR_BLOCK"`
